patches/nv_hooks.c: Rewrite runtime changes to the blob using signatures
The intention is to make these patches more durable, and therefore, make
updates easier. However, the blob base and size still must be inserted
manually at this time.

Performance is expected to be lower, as we must search for a handful
to dozens of signatures within the blob. Future optimisation is possible,
but DRAM is fairly fast and so the wait will be bareable to users.
However, it stalls the kernel for long enough that later versions/
pre-emptible builds complain.

Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>